Handle uncaught internal errors in device_pool
Also reorder the error scopes such that we catch OOM in the inner
most scope. We also check for OOM first. This is because it does
not allow device reuse. Errors that result in no reuse should be
checked first. Otherwise, if there is both a validation error and
an OOM error, and we check for validation first, we may end up
throwing the error and reusing the device because we didn't check
for OOM as well.