Add missing toStorage() when assigning values to outputs
This patch adds the missing call of toStorage() when assigning values
to outputs so that values can be correctly transformed to be assigned
to outputs.value.

This patch fixes the bugs when values are boolean values. In this case
outputs.value are uint values, so the missing of toStorage() on values
will cause WGSL compilation errors about type mismatch.

With this patch the below CTS tests can pass:
- webgpu:shader,execution,expression,binary,bool_logical:*