Compat: verify compressed format copies by rendering
The image_copy tests and copyTextureToTexture tests were using
`copyTextureToBuffer` to validate `writeTexture`,
`copyBufferToTexture`, and `copyTextureToTexture` but
`copyTextureToBuffer` does not work in compat for compressed
texture formats.

So instead, for those formats, create one texture and update
a portion of its content with `writeTexture`, `copyBufferToTexture`,
or `copyTextureToTexture`

Create another texture, this texture is the same size and format but
has its content generated on the CPU. In other words, the equivalent
of `writeTexture` / `copyBufferToTexture` / `copyTextureToTexture`
is executed on a Uint8Array that represents the entire content of the
texture.

The 2 textures are then rendered, layer by layer, to an rgba8unorm
texture and each render is read to a buffer. The buffers from
each texture's layers are expected to match.