Validate ptr type spelling for lets and params of user-declared functions
Check combinations of address space and access mode.
Check that we can read from any pointer.
Check that we can write to pointers that we should be able to, including
when the let infers the pointer type.

The store type of a pointer must be instantiable as the store type
of a variable that could be declared somewhere.

A function formal parameter must not be pointer to workgroup,
where the store type is atomic or contains an atomic.

Fixes: #2832 #1427 #1589