Reject invalid test params (#2862)
* Reject invalid test params

This PR rejects test params who's values are functions
or non-plain JavaScript objects.

This affects
  src/unittests/floating_point.spec.ts (toVector, toMatrix)
  src/unittests/texture_ok.spec.ts (findFailedPixels)
  the test mentioned in https://github.com/gpuweb/cts/issues/2860

This validates both case params and subcase params. I thought about
not validating subcase params because they are not part of the public
interface (they don't appear in URLs nor expectations) but, when
debugger it's common to comment out `beginSubcases`. This makes
it easy to run and debug individual subcases as cases. But, if
subcases are allowed to break these rules then the ability to quickly
comment out `beginSubcases` would not be possible.