Change texture/float comparisons for Inf
The texture comparison code needs to handle
`Positive.Infinity === Positive.Infinity` and
`Negative.Inifitiy === Negative.Infinity` which
needs an explicit check as `a - b`
returns `NaN` when a and b are infinity.

Further, the we need to properly encode/decode
inifinity and NaN in order for the comparisons
to round trip.