Remove orientation from several texture copy tests (#2914)
* Remove orientation from several texture copy tests

The orientation argument for these tests was copied from tests which
use an ImageBitmap source. Because CreateImageBitmap has an
imageOrientation option it made sense to test against all variants
of it. However, other tests, such as those that consume ImageData
directly, don't have any native mechanism for Y-flipping the source.
The data could be generated Y-flipped, but that doesn't increase
coverage of any implementation features. Despite not flipping the
source orientation, however, the test was still expected the data
to be flipped.

By removing this argument and no longer testing for flipped source
data, these tests all begin passing with no loss of platform coverage.

Bug: dawn:2017

* Lint fix