Split up 'scalar' tests into 'scalar' and 'vector' tests
The existing code was using vectorize to generate vector subcases of
the 'scalar' fixture, which lead to a confusing situation where the
'scalar' tests were testing scalar-scalar inputs and vector-vector
inputs, but not matrix-matrix, scalar-vector, or vector-scalar.

This PR factors out the vector-vector cases into their own fixture
named 'vector', but still retains the usage of vectorize to reduce the
need to test case duplication. Beyond clarity, this also divides up
the existing tests into smaller chunks to help with load-balancing,
etc.

Fixes #2935