Compat: Limit draw test attributes in compat
In compat mode, @builtin(vertex_index) and @builtin(instance_index)
each take an attribute so adjust this test to account for that.

I don't know if this test should be refactored to make sure
16 attributes are tested (So without using the builtins).
My feeling is this is probably fine?

The test in
webgpu/compat/api/validation/encoding/render_pipeline/vertex_state.spec.ts
does a simple test that you can create a pipeline with maxVertexAttributes
but it does not actually draw with that pipeline.