Compat: Limit max attributes (#2953)
In compat @builtin(vertex_index) and @builtin(instance_index)
each take an attribute so account for that in this test.

It's possible we should refactor this test to not
use vertex_index, instance_instance. For example we could
make each pair of data generate the correct pixel position.

For now it seemed best to get it to pass.
The test in webgpu/compat/api/validation/encoding/render_pipeline/vertex_state.spec.ts
does a simple test that you can use maxVertexAttributes but it does
not test all the combinations like this test. It only tests that
creating a pipeline passes if you use maxVertexAttributes and
fails if you use maxVertexAttributes + the builtins above