Make Pass > Skip
As it is, the test recorder only keeps the "highest" status. It
was Pass = 0, Skip = 1, that means if 1000 subcases passes but just
one was skipped the entire case was marked as skip.

Making Pass = 1 and SKip = 0 means they'll as long as one subcase runs
the test will be marked as Pass unless there is higher a higher status
subcase