wgsl: Speed up `ldexp` execution tests
Reduces thh number of cases by using sparse instead of full ranges,
since there is going to be a cartesian product of input values when
generating cases.

Optimizes two quantization functions that had not been updated to
re-used their TypedArray. Creation and then immediate destruction of
TypedArrays are a type of hotspot we have encountered in other areas
of the code base.