Use largest maxInterStageShaderVariables in maxInterStageShaderComponents tests
This patch uses the largest value of maxInterStageShaderVariables supported on
current adapter in the tests about maxInterStageShaderComponents when
creating devices so that when the value we use as maxInterStageShaderComponents
is larger than the default one, we won't be limited by the default value of
maxInterStageShaderVariables.

This patch also removes the assertion that the value of
maxInterStageShaderVariables must be larger than a quarter of
maxInterStageShaderComponents as on many backends the largest value of
maxInterStageShaderComponents is equal to 4x maxInterStageShaderVaraibles, so
in "overLimit" tests the value of maxInterStageShaderComponents can be greater
than 4x device.limits.maxInterStageShaderVaraibles.