wgsl: Convert `quantizeToF32` to used `Math.fround`
Instead of passing the input through a F32Array, use the builtin
Math.fround.

This leads to a ~5% improvement benchmarking locally. This is less
than the equivalent f16 change, because F32Array is provided by the
runtime, whereas F16Array is being polyfilled, so is probably more
efficient to begin with.