Speed up shader,execution,expression tests with async pipelines
Some of these tests are among the longest running that we have, so
hopefully this has a dramatic impact on runtimes. Profiling shows
that almost all the time is being spent waiting for pipeline
creation. By using createComputePipelineAsync, however, we ensure
that multiple shaders in a batch can be compiling at once, which
drops test run times significantly.