Remove preview workflow, do WPT lint in PR workflow
The artifact upload for the PR preview workflow was really slow. We
could re-enable this if we find it's needed though.

The WPT lint workflow was doing a redundant build of most of the CTS. We
can just run WPT's lint after a regular build (it seems to take only 2
seconds on past runs).