Test writeTimestamp never works
writeTimestamp has been removed from the spec so it should always
TypeError. (Technically it shouldn't exist at all but per our error
design this is sort of OK since it's the same thing you would get if
writeTimestamp were behind a feature that's not enabled.)

Passes on Chrome Canary, confirmed fails after enabling
chrome://flags/#enable-webgpu-developer-features enabled (which causes
it not to throw a TypeError).