Remove compression on the case cache, split files to reduce transitive deps (#3197)
* Remove compression on the case cache

Compression is not stable between Node versions, nor between operating systems.

This inflates the cache size 10x (from 4.3M to 49M), however these files should not be frequently changing, and Git should be performing compressions with packfiles and during network operations.

* Split cache generation from spec.ts files

* Breakup monolithic imports

Move commonly used code out of monolithic imports to reduce the amount of code imported by the `.cache.ts` files.

Add a new `MUST_NOT_BE_IMPORTED_BY_DATA_CACHE` token that can be used to figure out why the cache wants to regenerate for a given file. Add this to device_pool.ts, as this really shouldn't be depended on for the cache.