Allow file paths as queries
This allows you to use a file  path as a query.
A path is a string that contains `/` or `\\` and
ends in `.spec.ts`. If a query looks like a path
then the `.spec.ts` is removed, the string is split
by `\` and `//`. Then part after the last 'src' is
converted to a query. So for example

/Users/me/src/gpuweb/cts/src/webgpu/foo/bar/moo.spec.ts ->
  webgpu:foo,bar,moo,*

C:\home\me\src\gpuweb\cts\src\webgpu\foo\bar\moo.spec.ts ->
  webgpu:foo,bar,moo,*

webgpu/foo/bar/moo.spec.ts ->
  webgpu:foo,bar,moo,*