Add support for AbstractInts as inputs (#3306)
This PR adds in abstract ints to the Scalar type system, and wires up
being able to write them as inputs as literals in shaders. Since they
are abstract values they will never be supported as buffer
inputs. Returning the values via buffers will be handled in a seperate
PR.

I have implemented the tests for AbstractInt -> i32 & u32 conversion
to confirm the implementation operates as expected.

There is a substantial number of changes related to the fact that the
existing code assumes all numerics are `number`s, but AbstractInts
need to be represented as BigInts.

I have also cleaned up a couple of places related to the AbstractFloat
implementation where there wasn't full testing coverage, etc.

Issue #3302