Fix assignment of AInt to u32 in the expression generator (#3379)
Apply the same fix that was made for AFloat to AInt when emitting the
value assignment WGSL. Otherwise, if
globalTestConfig.unrollConstEvalLoops is true, as it is on Windows, then
we emit invalid WGSL because we index an array of AInts with a runtime
index, forcing concretization, which results in an assignment of i32 to
u32. The fix is to do as was done with AFloat: we emit direct
assignments of the literal values into the output array. Also made sure
we don't emit the array of values in this case, as it's unused.