Separate tests for target format supporting usage and being a color format (#3393)
This splits two logically separate tests that were part of the same test before. This will make it easier to fix usage of the texture info table: Previously, it would access certain members of the info table assuming they were color, but would actually get the depth component of depth and depth-stencil formats. This wasn't really a problem because those formats were already expected to fail. But cleaner to instead just scope that test to only color formats.