Add compat mode testing to the PR checklist
I'm not sure it's valid to put this here as for example, Safari
will not have compat mode. But, at least at the moment, we have
the issue that new tests can fail in compat and should either be
refactored if they can be or else skipped.

It's also possible to enforce compat validation via an extension
so I could go write one for Safari if the need is there.