Add AdapterLimitsGPUTest that sets all limits to the adapter's
Note: I'm not sure how I refactored device_pool.ts is good
but, the old code created one initial device with no descriptor
the first time in acquire. It then created another of the type
actually wanted by the user. That no longer works becauase the
adapter is passed in and so if the first requestDevice succeeds,
then the adapter has been used and can't be used for the second
one.

Also, assuming this change is approved,  I can (and should?) probably
refactor LimitsTest to use this vs how it's doing it now. That
would end up using cached devices.