Fix device_pool refactor (#3469)
The old version made one device with no descriptor.
If that fail then it stopped creating any devices period.

The previous refactor was bad in that if the any device
failed it would just stop making devices.

This one is simplified. It just used the pool as a pool.
There is no special handling for failed devices.