Clean up deprecated .copyDst/.renderable/.renderTarget* from kTextureFormatInfo (#3412)
* Refactor deprecated .copyDst/.renderable in texture_zero_init_test.ts

* Remove no-op use of .renderable in texture_zero_init_test.ts

* Refactor deprecated .renderable in createTexture.spec.ts

* Remove deprecated .copyDst

* Remove deprecated .renderable

* Remove unused .renderTarget{ComponentAlignment,PixelByteCost}

.renderTargetComponentAlignment = .colorRender.alignment
.renderTargetPixelByteCost = .colorRender.byteCost

The one exception to this is rg11b10ufloat, which is not renderable in
core, but has an extension for renderability. kTextureFormatInfo
currently only expresses core capabilities. The alignment and byteCost
for this format will need to be added back when the refactor is done to
make a table for "static" properties of texture formats (which never
vary based on capabilities) - shape of that refactor is TBD.

* Refactor trivial uses of deprecated .bytesPerBlock