shader/execution: Fix struct padding for `f16`s
`wgslMembers()` was using an `i32` to pad the structure to its alignment, however `i32` has an alignment and size of 4 bytes.
If the structure holds an odd number of `f16` scalars, then `i32` cannot be used to account fo this padding due to the 4-byte size and alignment.
Pad with `f16` if the padding is an odd multiple of 2 bytes.