Fix depth_clamp_and_clip test (#3524)
Broken in #3412. Prior to this, some copies were being skipped for some formats because bytesPerBlock was undefined for those copies.
This doesn't change the coverage of the test though, as the actual check that used those copies is still skipped.

I missed the regressions before because they were a bit buried in existing failures on my machine.