Use -1000 to +1000 for float formats
This allows the tests to pass on Swiftshader. It's also close to
the values used by the OpenGLES CTS. It uses +/-1000 for F16
and +/-100000 for F32.

Note: With this change, finiteMin and finiteMax are unused but
I'd expect them to be used for `textureLoad` to test round trip
values through shaders.