shader/execution: Fix abstract non_zero constructor tests
Depending on the platform, we either directly assign the expression evaluation to the output buffer, or go via a const array.
The latter caused compilation failures as it would force a concretization of the abstract values in the array, which would become incompatible with the output type.

To force the harness to avoid the array, add a new optional `ConstEvaluationMode` flag to force the evaluation mode used.

Also replace the growing list of `ShaderBuilder` parameters with a structure, so that fewer places need to be updated for this sort of change in the future.

Issue: #3657