shader/exeuction: Fix abstract-numeric constructor tests
Specify `constEvaluationMode: 'direct'` for abstract-numeric constructor tests.

This CL extends #3677 to fix more tests that early concretize values when loops are not unrolled.

Depending on the platform, we either directly assign the expression evaluation to the output buffer, or go via a const array. The latter caused compilation failures as it would force a concretization of the abstract values in the array, which would become incompatible with the output type.

Fixes: #3743