WGSL builtin textureSample w/derivatives test (#3757)
* WGSL builtin textureSample w/derivatives test

This test is different than the pattern set by the non-derivative
test. The existing test renders an array of coordinates. The coordinates
might come from data passed to the shader, or they might be hard coded
with multiple `textureSample` calls. The results are written to a
storage buffer and compared to computed expected values.

This test instead renders a quad (2 triangles) to an rgba32float texture
given a simple shader. Then "software renders" the expected result
and compares the 2 results.

There are a large number of permutations of parameters that affect
texture sampling. The sampler itself has wrapping modes and sampling
modes. `textureSample` is influenced by texture format, coordinates,
derivatives and offsets. Testing all combinations X all texture formats
would make the tests too slow so, these test focus mostly on the effect
of derivatives and assume the combinations in the non-derivative case
give effective coverage of the other permutations.

The tolerances for whether the actual result match the expected result
are rather large. They are influenced by the OpenGL dEQP test suite
which has similar tolerances. The hope is these are a good starting
point to get the tests written with a structure that can be adjusted
later should it be decided to make these tests more strict.