WGSL builtin textureSample compressed formats
I'm not sure these are needed or not but given we were testing
so many other formats it seems like we should be testing
compressed formats?

It does this by filling a compressed texture with random data.
For compressed textures of type unorm or snorm this is fine.
For float types this would fail since they random values might
be large, Infinity, or NaN. We aren't converting from values
to a compressed format. We're filling a buffer with binary
random data and assuming the values that come out are in range.

After that, read that texture back to the GPU using a compute
shader. This way we get WebGPU to decode the compressed texture
to an uncompressed format.

Some places I got lost. I first put in the derivatives test
and it passed. Then I put in non-derivatives test and it failed
on srgb formats. This makes some sense I think because I was
using only `rgba32float` as the format for pulling the data out
of the texture.

That meant, when doing the software interpolation it was probably
doing it wrong. I switched to using one of `rgba8unorm`,
`rgba8snorm` or `rgba8unorm-srgb` as the format the texture is
read back to `TexelView` as. I think ideally if I could have an
`rgba32float-srgb` that would have been best but that format
doesn't exist and I think hacking it into `TexelView` would have
made `TexelView`'s list of formats not a subset of WebGPU's

Also, I'm pretty worried about the number of tests. Just these 4
tests take 42 seconds to run on my M1 Mac and there's still
22 tests to go for this single built in alone and then all the rest
of the texture builtins.