Rework test case generation for determinant test
1.  Constrain matrix element sizes for determinant test.
Previously, the f32 inputs were being used for f16, but that can lead
to excessive loss of accuracy, and trigger spurious failures.

The cases for the determinant tests should be easily computed
without losing accuracy.

The matrix elements will have a fixed budget for significant bits.
That budget should be a function of the significand size for the
floating point type, and the dimensions of the matrix.

For f16, there are surprisingly few bits available.
   4x4 => 2 bits (!)
   3x3 => 3 bits
   2x2 => 4 bits

For f32, many more bits can be budgeted:
   4x4 => 21/4
   3x3 => 22/3
   2x2 => 23/2

2.  Use randomly generated input matrices.  Previously, for 3x3 and
4x4 cases, the matrices would always have a determinant of 0.
(That's because two of the rows would be linearly dependent; a
scalar multiple of the other.  Similarly two columns would be
linearly dependent.)