refactor shader/types.ts: Add accessSuffixes to type info (#3864)
When testing accesses into uniform address space, the array
tests pack scalars into a 4-element vector.  This is a big hack,
to accommodate the 16byte alignment requirement for uniform address
space. Previously the robust access tests had a subtle tight coupling
to this decision.

This change adds a 'accessSuffixes' field to the array type info
to indicate how to access individual scalar elements under test.
So in the uniform case, when a vec4 is artificially created, return
a list of the .x .y .z .w components to get the original scalar
values back.  This allows client tests to smoothly generalize.

Update the robust access tests to this change. It's the only affected test.

Issue: #3405