Add dual source blending blend factors to `kBlendFactors`
This patch adds all the blend factors defined in the extension
"dual-source-blending" to `kBlendFactors` in capability_info.ts so
that all the tests about blend factors can test these newly added
blend factors.

This patch also removes the validation tests about using the alpha
channel of `Src1` because it doesn't follow the latest SPEC: only when
the color blend factor uses the alpha channel of `Src` must the fragment
output be `vec4`. When the color target doesn't have the alpha channel,
the alpha blend factor is ignored so it is OK to set it as if it "uses"
the alpha channel of `Src`.

Note that the new validation test against `Src1-alpha` will be added into
'pipeline_output_targets,blend' after the fix is landed in the browser.