Add validation tests on the access of the alpha channel of Src1
This patch adds validation tests on the use of `src1-alpha` in the
blend factor. When the blend factor uses the alpha channel of src1,
the fragment output must be vec4.

This patch also implements function `IsDualSourceBlendingFactor()`
to check if a blend factor needs the extension "dual-source-blending".

This patch also adds the support of dual source blending in the helper
function `getFragmentShaderCodeWithOutput()` to simplify the shaders
used in dual source blending tests.