Only report on first texture builtin error
Each texture test tests 50 calls to `textureXXX`. It then
checks they all match as expected. For each one that does NOT match
expected a slow binary search happens to find the samples that
influenced the results.

This change makes it so only the first failure does this slow
check. Ideally no tests would fail but when they do, checking all
50 sample points can cause test bots to timeout.