Use the same pipeline for identifySamplePoints
Before this change, the code would generate one pipeline
for checking a bunch of texture builtin calls (50). If one of
them failed it would generate a new shader with just the single
call that failed as well as new inputs for that single call and
then use that to figure out the sample points.

After this change is just calls the same existing pipeline.
This means if the error is because of the larger 50 call pipeline
it should produce the same results whereas before, with a different,
simpler pipeline, it might not have produced the same results.

Further, not creating a new pipeline and new inputs is faster.

With this change,