Speed up Texture Builtin tests
The slowest part of the test is generating random textures via
`TexelView.fromTexelsAsColors`. This is because it calls a deep
hierarchy of functions to generate a PerTexelComponent<number> per
texel where the values are quantized to match what the GPU should
return so that the software texture functions works with similar values.

For many formats though, we can just fill a typearray with random data
and pass that data to `TexelView.fromTextureDataByReference` which is
orders of magnitude faster.