Fix textureSampleBias
The default gradient mult used [1,1,1] and then set one
of them the value that would generate the required mipLevel.
But that was ignoring the that if the required mipLevel is
less then 0 the 1s left over would generate a mipLevel of 0.

For now, make the default [0,0,0]. Another alternative
would be to set them all to the desired value that generates
the desired mipLevel or, we could set them to something less
than the value that generates the desired mip level.