Update WebGPU types to 0.1.50
This makes toneMapping optional which required updating a few tests to
use the nullish coalescing operator when accessing it.

New vertex formats are added that require updating capability_info.ts
(and tests for unorm8x4-bgra is added at the same time).

Finally dictionary types for copies are renamed throughout the CTS even
if deprecated aliases are available.