Fix readTextureToTextureViews for multisampled textures
I'm not sure when this bug was introduced. These tests used
to pass. Fixed the bug and Added a test to hopefully find it
quicker if it breaks in the future.

Also added a test for createTextureWithRandomDataAndGetTexels
with a generator. This was fixed in a previous PR without a
a specific test as the fact that it failed other tests showed
it needed fixing but I decided it's probably best to keep
these tests too.