Fix readTextureToTextureViews for multisampled textures (#4049)
* Fix readTextureToTextureViews for multisampled textures

I'm not sure when this bug was introduced. These tests (textureLoad:multisampled) used
to pass. Fixed the bug and added a test to hopefully find it
quicker if it breaks in the future.

Also added a test for createTextureWithRandomDataAndGetTexels
with a generator. This was fixed in a previous PR without a
a specific test as the fact that it failed other tests showed
it needed fixing but I decided it's probably best to keep
these tests too.

* Update src/webgpu/shader/execution/expression/call/builtin/texture_utils.ts

Co-authored-by: Kai Ninomiya <kainino1@gmail.com>