Test texture mix weights separately (#4051)
* Test texture mix weights separately

Before this change, the texture builtin utils would query
how the local GPU mixes texture mip levels. It expected
that to be close to a linear interpolation which is true
for most GPUs. The known exceptions are Mac AMD and Mac
Intel in a compute stage.

It would check how far from linear iterpolation the weights
are. If the weights were too far off linear interpolation
it would assert.

With this case, those asserts for linear interpolation are
moved from all the tests that use these weights to a separate
test.

This allows the texture builtin tests to run further on devices
that have bad mix weights with the new test at least documenting
a particular GPU's failure to implement linear interpolation
between mip levels.