Add MaxLimitsTest that sets all limits to the adapter's
This is what I came up with. Without a major refactor the
issue is the adapter is requested at an extremely low level
but a test needs access to an adapter to select limits.

So, made selectDeviceOrSkipTestCase pass an optional callback
all the way down that gets passed the adapter and at
CanonicalDeviceDescriptor giving it a chance to modify it
down at the level where an adapter is available.

More of the CTS tests need to be testing with max limits.

Issue: #3363