Move MaxLimitsTest to MaxLimitTestMixin (#4085)
I wanted to use MaxLimitsTest with resource_compatibility.spec.ts
but that file's tests are based on `CreateRenderPipelineValidationTest`
which itself is based on `ValidationTest` and so would force making
`ValidationTest` be based on `MaxLimitsTest`. I actually think thats
a good idea. The default should be max limits. But, it breaks too
many things.

So, MaxLimitsTest a mixin.