Fix DeviceModifier
The issue is the key is selected before the modification happens
which means the devices in the pool will not match the key.

I can't think of a way to fix this using the adaptor so maybe
this hacky solution works by letting you modify the key.

Generally you just append a string but I made it function because
MaxLimitsTestMixin wants to be able to chain to other modifiers.

Maybe it should all be re-designed

This came up because I notice a test passing that shouldn't have
passed. It passsed because a `MaxLimitsTestMixin` test ran first
and it's key was `''` which meant that other tests got a max limits
device from the pool.