Compat: Add tests for maxStorage(Buffer/Texture)sIn(Vertex/Fragment)Stage
Note: I couldn't really think of a test on the core side.
Originally I was going to try to request `maxXXXInYYYStage` less
than `maxXXXPerShaderStage` which is what the tests do in compat
but in core I believe it will be a no-op and that InStage limits
will always match their PerStage counterparts.

I added a `RequiredLimitsTestMixin` to try to make it easier to
write a test that needs specific limits. The issue is the key.
I thought about making the key just be the source of `getRequiredLimits`
but I can imagine someone writing some more generic requester
that takes a list of of limits in which case the code would be generic
and so using the source of the funciton would not be a valid key.

I also though about adding a MaxOfSpecificLimitsMixin where you could
pass in a list of maxes but I think most code that needs maxes can use
MaxLimitsTestMixin